Introduced defensive programming
There were a few spots where somebody could enter input which would break the program. Have gone through and cleaned the more noticeable of the bunch.
Including when asking for an input for the key, and the input for the cleartext. As if somebody entered a special character (expect spaces) or a number it would break.
Also broke down the input for cleartext into a separate function to be usable for the decryption process as well.